requirements.txt
Django==4.1.1
psycopg2>=2.8
django-jazzmin==2.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Markdown==3.4.1
django-filter==22.1
django-jazzmin==2.5.0
django-cors-headers==3.6.0